depend on `flutter_map_dragmarker` package
